EWPP-1410: Move related code realted to default author value to related submodule.
