+++ oe_content.module
@@ -118,7 +118,7 @@ function oe_content_node_prepare_form(EntityInterface $entity, $operation, FormS
     return;
   }
 
-  if ($entity->hasField('oe_content_content_owner') === FALSE && $entity->hasField('oe_authors') === FALSE) {
+  if ($entity->hasField('oe_content_content_owner') === FALSE) {
     return;
   }
 
@@ -132,18 +132,6 @@ function oe_content_node_prepare_form(EntityInterface $entity, $operation, FormS
     // No content owner(s) defined.
     return;
   }
-
   $content_owners = $site_information->getDefaultContentOwners();
-
-  if ($entity->hasField('oe_content_content_owner')) {
-    $entity->set('oe_content_content_owner', array_values($content_owners));
-  }
-  // Set default corporate body entity for authors field.
-  if ($entity->hasField('oe_authors')) {
-    $corporate_body = \Drupal::entityTypeManager()->getStorage('oe_author')->create([
-      'type' => 'oe_corporate_body',
-      'oe_skos_reference' => array_values($content_owners),
-    ]);
-    $entity->set('oe_authors', [$corporate_body]);
-  }
+  $entity->set('oe_content_content_owner', array_values($content_owners));
 }
